+++ packages/types/src/BaseModelCard.ts
